Fixed according to the code review
