+++ benchmarks/scripts/generate_plots_flow_flatten_merge.py
@@ -3,9 +3,9 @@ import matplotlib.pyplot as plt
 from matplotlib.ticker import FormatStrFormatter
 
 inputFile = "jmh-result.csv"
-outputFile = "flatten-merge-plots.svg"
-elements = 10000
-benchmarkName="benchmarks.flow.FlattenMergeBenchmark.flattenMerge"
+outputFile = "flow-flatten-merge.svg"
+elements = 100000
+benchmarkName="benchmarks.flow.FlowFlattenMergeBenchmark.flattenMerge"
 
 markers = ['.', 'v', '^', '1', '2', '8', 'p', 'P', 'x', 'D', 'd', 's']
 colours = ['black', 'silver', 'red', 'gold', 'sienna', 'olivedrab', 'lightseagreen', 'navy', 'blue', 'm', 'crimson', 'yellow', 'orangered', 'slateblue', 'aqua']
@@ -29,7 +29,7 @@ def draw(data, plt):
     plt.gca().xaxis.set_major_formatter(FormatStrFormatter('%0.f'))
     plt.grid(linewidth='0.5', color='lightgray')
     plt.ylabel(data.unit.unique()[0])
-    plt.xlabel('parallelism')
+    plt.xlabel('concurrency')
     plt.xticks(data.concurrency.unique())
 
     colourGen = next_colour()
@@ -39,13 +39,11 @@ def draw(data, plt):
         genMarker = next(markerGen)
         res = data[(data.flows == flows)]
         plt.plot(res.concurrency, res.score*elements, label="flows={}".format(flows), color=genColour, marker=genMarker)
+        plt.errorbar(x=res.concurrency, y=res.score*elements, yerr=res.scoreError*elements, solid_capstyle='projecting', capsize=5)
 
-def genFile():
-    data = pd.read_table(inputFile, sep=",", skiprows=1, names=["benchmark","mode","threads","samples","score","scoreError","unit","concurrency","flows"])
-    plt.figure(figsize=(20, 20))
-    draw(data, plt)
-    plt.legend(loc='upper center', borderpad=0, ncol=4, frameon=False, borderaxespad=4, prop={'size': 8})
-    plt.tight_layout(pad=12, w_pad=2, h_pad=1)
-    plt.savefig(outputFile, bbox_inches='tight')
-
-genFile()
+data = pd.read_table(inputFile, sep=",", skiprows=1, names=["benchmark","mode","threads","samples","score","scoreError","unit","concurrency","flows"])
+plt.figure(figsize=(20, 20))
+draw(data, plt)
+plt.legend(loc='upper center', borderpad=0, ncol=4, frameon=False, borderaxespad=4, prop={'size': 8})
+plt.tight_layout(pad=12, w_pad=2, h_pad=1)
+plt.savefig(outputFile, bbox_inches='tight')
